refactor(deployments): rename app-card to deployment-card
related to openshiftio/openshift.io#1382
