+++ src/app/space/create/apps/apps.component.html
@@ -34,7 +34,7 @@
             {{application}}
           </h4>
         </div>
-        <app-card class="col-xs-6 col-sm-4 col-md-3" [applicationId]="application" [environment]="environment" *ngFor="let environment of environments | async"></app-card>
+        <deployment-card class="col-xs-6 col-sm-4 col-md-3" [applicationId]="application" [environment]="environment" *ngFor="let environment of environments | async"></deployment-card>
       </div>
     </div>
   </div>
